Now Mozart compile mostly warning-free again! + Fixed most warnings about missing const with literal strings (remaining one is for virtualString to CString which is to be reviewed) + Changed the reporting flags in the DSS to sane values (removed -Weffc++ and -Winline, forced the use of most flags for GCC > 3) + Fixed most of the remaining warnings but   - usage of tmpnam -> requires some user-visible changes   - warnings in ozbison; well it's in code that's more than 12(!) years old. The good way would be to update the underlying bison. Pain ahead. + Fixed some bugs here and there (warnings are supposed to help locate them when they aren't drown in noise) +The output of make bootstrap dropped from 1.7 Mb to 189 kb!
git-svn-id: https://gforge.info.ucl.ac.be/svn/mozart@17178 ada56829-ad1f-0410-b00f-83cda6628aec

+++ trunk/mozart/platform/emulator/libdp/glue_buffer.hh
@@ -33,8 +33,8 @@ class GlueReadBuffer:public ByteBuffer,
   
 public: 
   GlueReadBuffer(BYTE*, int); 
-
-  virtual int availableData() const; 
+  
+  virtual size_t availableData() const; 
   virtual bool canRead(size_t len) const {return this->availableData()>=len;}
   virtual void readFromBuffer(BYTE* ptr, size_t wanted);
   virtual void commitRead(size_t read);
@@ -49,7 +49,7 @@ class GlueWriteBuffer:public ByteBuffer,
 public: 
   GlueWriteBuffer(BYTE*, int); 
 
-  virtual int availableSpace() const; 
+  virtual size_t availableSpace() const; 
   virtual bool canWrite(size_t len) const {return this->availableSpace()>=len;}
   virtual void writeToBuffer(const BYTE* ptr, size_t write); 
 
@@ -70,14 +70,14 @@ public:
   ~GlueMarshalerBuffer() {}
 
   // DssReadBuffer interface
-  virtual int availableData() const { return 1000000; }     // ahem...
+  virtual size_t availableData() const { return 1000000; }     // ahem...
   virtual bool canRead(size_t) const { return true; }
   virtual void readFromBuffer(BYTE*, size_t);
   virtual void commitRead(size_t) {}
   virtual const BYTE getByte() { return buffer->get(); }
 
   // DssWriteBuffer interface
-  virtual int availableSpace() const { return 1000000; }     // ahem...
+  virtual size_t availableSpace() const { return 1000000; }     // ahem...
   virtual bool canWrite(size_t) const { return true; }
   virtual void writeToBuffer(const BYTE* ptr, size_t write);
   virtual void putByte(const BYTE& b) { buffer->put(b); }
